Restructure the code, add interfaces
- Fix stupid old coding design with added namespace depths.
- Changed stupid plural form to singular for exceptions.
- Added interfaces to the main objects.
