+++ src/Exception/InvalidCallException.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace hisorange\BrowserDetect\Exception;
+
+// PHP global.
+use Exception;
+
+/**
+ * Class InvalidCallException
+ *
+ * @package hisorange\BrowserDetect\Exception
+ */
+class InvalidCallException extends Exception
+{
+}
