+++ src/Exception/InvalidPluginListException.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace hisorange\BrowserDetect\Exception;
+
+// PHP global.
+use Exception;
+
+/**
+ * Class InvalidPluginListException
+ * @package hisorange\BrowserDetect\Exception
+ */
+class InvalidPluginListException extends Exception
+{
+}
