--- src/Exceptions/InvalidPluginListException.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace hisorange\BrowserDetect\Exceptions;
-
-// PHP global.
-use Exception;
-
-class InvalidPluginListException extends Exception
-{
-}
